refactor: move artifact command into a separate module
